refactor: split spec and impl of filter state and filters
