+++ helper_dart/lib/algolia_helper.dart
@@ -12,7 +12,6 @@ export 'src/filter_state.dart';
 export 'src/filters.dart';
 export 'src/highlighting.dart';
 export 'src/hits_searcher.dart';
-export 'src/immutable_filters.dart';
 export 'src/search_response.dart';
 export 'src/search_state.dart';
 export 'src/selectable_item.dart';
